unitedkingdom.txt
This is the raw data for the United Kingdom example.  In the Simple version of AstroGIS, the Polycreate program is run to create a coverage file from this data that the program can read as a module.  The program created from this data, updated only slightly with some additional default variable values, is included in the "Library" subfolder.